# Faster compilation of functions