This change includes caching and smarter analysis of classpaths
to limit the amount of jar walking necessary to find dependencies
when compiling. It also includes some new tests that verify
packaging of dependencies in boot style form (BOOT-INF/lib and
BOOT-INF/classes).

It also includes some fixes (and tests) for resolution of types in
the default package.

A simple speed test is included, if I run this on master (without these changes) it reports:

Reusing FunctionCompiler: #1 = 3077ms
Reusing FunctionCompiler: #2 = 2297ms
Reusing FunctionCompiler: #3 = 2146ms
Reusing FunctionCompiler: #4 = 2162ms
Reusing FunctionCompiler: #5 = 2205ms
New FunctionCompiler: 2162ms
New FunctionCompiler: 2108ms
New FunctionCompiler: 2149ms

With these changes applied:
Reusing FunctionCompiler: #1 = 517ms
Reusing FunctionCompiler: #2 = 392ms
Reusing FunctionCompiler: #3 = 378ms
Reusing FunctionCompiler: #4 = 467ms
Reusing FunctionCompiler: #5 = 461ms
New FunctionCompiler: 362ms
New FunctionCompiler: 368ms
New FunctionCompiler: 355ms